requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
gunicorn==22.0.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
Flask-PyMongo==2.3.0
pymongo==4.11.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
requests==2.32.3
